Use latest ``python3.x`` version for ``manylinux2014``
We do not need to distinguish between ``python3.x`` and ``python3.xm``, since
the two interpreters are identical for Python version >=3.6. Therefore, we
switch to using

  PYTHON = /usr/local/bin/python3.x

instead of the specific Python version (e.g., 3.6.14). This avoids having to
update the Dockerfile everytime the source image is updated to a new Python
version.